복사생성자 금지 매크로에 Clone() 금지 추가

+++ Player.cpp
@@ -61,7 +61,7 @@ void Player::Init()
 	AddSkillKeyMap(KEY_CODE::SHIFT, *activeSkill);
 
 	// 자식 오브젝트 추가
-	SkillObject* skillObject = (SkillObject*)AddChild(SkillObject(L"", Vec2<float>(scale.x, 0.f), Vec2(20, 20), LAYER_TYPE::PLAYER_SKILL));
+	SkillObject* skillObject = (SkillObject*)AddChild(SkillObject(L"", Vec2<float>((float)scale.x, 0.f), Vec2(20, 20), LAYER_TYPE::PLAYER_SKILL));
 	((AttackActiveSkill*)activeSkill)->SetSkillObject(*skillObject);
 
 	// 필드 초기화
